+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+- Made `applySnapshot` not check by model instance in reconciliation for better compatibility with hot-reloading.
+
 ## 0.36.1
 
 - Fixed `ObjectMap.forEach` and `ArraySet.forEach` typings.